+++ nexus/pkg/artifacts/manifest.go
@@ -1,12 +1,10 @@
 package artifacts
 
 import (
-	"crypto/md5"
 	"fmt"
 	"io"
 	"net/http"
 	"os"
-	"sort"
 
 	"github.com/segmentio/encoding/json"
 
@@ -64,33 +62,6 @@ func NewManifestFromProto(proto *service.ArtifactManifest) (Manifest, error) {
 		}
 	}
 	return manifest, nil
-}
-
-func (m *Manifest) ComputeDigest() string {
-	type hashedEntry struct {
-		name   string
-		digest string
-	}
-
-	var entries []hashedEntry
-	for name, entry := range m.Contents {
-		entries = append(entries, hashedEntry{
-			name:   name,
-			digest: entry.Digest,
-		})
-	}
-
-	sort.Slice(entries, func(i, j int) bool {
-		return entries[i].name < entries[j].name
-	})
-
-	hasher := md5.New()
-	hasher.Write([]byte("wandb-artifact-manifest-v1\n"))
-	for _, entry := range entries {
-		hasher.Write([]byte(fmt.Sprintf("%s:%s\n", entry.name, entry.digest)))
-	}
-
-	return utils.EncodeBytesAsHex(hasher.Sum(nil))
 }
 
 func (m *Manifest) WriteToFile() (filename string, digest string, rerr error) {
